# src/main/java/com/library/managementservice/api/dto/LoanDetailResponse.java
package com.library.managementservice.api.dto;

import java.time.OffsetDateTime;

public record LoanDetailResponse(
    Long loanId,
    Long bookId,
    String bookTitle,
    OffsetDateTime borrowedAt,
    OffsetDateTime dueDate,
    OffsetDateTime returnedAt
) {}